Dev

See merge request bms/scmp/scmp-api!167

+++ src/main/java/com/aiqin/bms/scmp/api/supplier/domain/request/supplier/dto/SupplyCompanyDetailDTO.java
@@ -94,7 +94,7 @@ public class SupplyCompanyDetailDTO {
     private String corporateRepresentative;
 
     @ApiModelProperty("注册资金")
-    private Long registeredCapital;
+    private BigDecimal registeredCapital;
 
     @ApiModelProperty("是否禁用")
     private Byte enable;
